Added a weird env-variable fix
Shinyserver sandboxes each app, which makes it a bit difficult to
pass env variables to let the apps know how to reach the database,
etc.

I did a dirty fix which just dumps the env into the cwd of the app.
This is potentially unsafe, but since the apps don't expose any
kind of dangerous input, i am letting it fly for now.